+++ src/features/amUI/altinn2Account/AddAltinn2AccountPage.tsx
@@ -8,9 +8,7 @@ import {
   DsHeading,
   DsParagraph,
   DsSpinner,
-  DsTextfield,
 } from '@altinn/altinn-components';
-import { PersonCircleIcon } from '@navikt/aksel-icons';
 import { useDocumentTitle } from '@/resources/hooks/useDocumentTitle';
 import { getAfUrl } from '@/resources/utils/pathUtils';
 import { useGetReporteeQuery } from '@/rtk/features/userInfoApi';
@@ -21,14 +19,14 @@ import {
 } from '@/rtk/features/selfIdentifiedUserApi';
 import classes from './AddAltinn2AccountPage.module.css';
 import { PageLayoutWrapper } from '../common/PageLayoutWrapper';
+import { Altinn2AccountForm } from './Altinn2AccountForm';
 
 export const AddAltinn2AccountPage = () => {
   const { t } = useTranslation();
   const [searchParams] = useSearchParams();
   const token = searchParams.get('token') ?? '';
-  const [userName, setUserName] = useState('');
-  const [password, setPassword] = useState('');
-  const [step, setStep] = useState<number>(1);
+
+  const [step, setStep] = useState<1 | 2 | 3 | 4>(1);
   const modalRef = useRef<HTMLDialogElement>(null);
 
   useDocumentTitle(t('add_altinn2_account_page.page_title'));
@@ -52,30 +50,24 @@ export const AddAltinn2AccountPage = () => {
     { isLoading: isAddingAltinn2AccountFromToken, error: addUserFromTokenError },
   ] = useAddAltinn2AccountFromTokenMutation();
 
-  const isValidateCrentialsButtonDisabled = !userName || !password || isAddingAltinn2Account;
-  const isSendEmailButtonDisabled = !userName || isSendingForgotPasswordEmail;
   const afUrl = `${getAfUrl()}inbox`;
   const profileUrl = `${getAfUrl()}profile`;
 
-  const onAddAltinn2Account = async () => {
-    if (!isValidateCrentialsButtonDisabled) {
-      try {
-        await addAltinn2Account({ userName, password }).unwrap();
-        setStep(3);
-      } catch {
-        // error displayed via addUserError RTK Query state
-      }
+  const onAddAltinn2Account = async (userName: string, password: string) => {
+    try {
+      await addAltinn2Account({ userName, password }).unwrap();
+      setStep(3);
+    } catch {
+      // error displayed via addUserError RTK Query state
     }
   };
 
-  const onSendForgotPasswordEmail = async () => {
-    if (!isSendEmailButtonDisabled) {
-      try {
-        await sendForgotPasswordEmail({ userName }).unwrap();
-        setStep(5);
-      } catch {
-        // error displayed via forgotPasswordError RTK Query state
-      }
+  const onSendForgotPasswordEmail = async (userName: string) => {
+    try {
+      await sendForgotPasswordEmail({ userName }).unwrap();
+      setStep(4);
+    } catch {
+      // error displayed via forgotPasswordError RTK Query state
     }
   };
 
@@ -97,22 +89,6 @@ export const AddAltinn2AccountPage = () => {
       modalRef.current?.showModal();
     }
   }, [reportee, token]);
-
-  const handleInputFieldKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      onAddAltinn2Account();
-    }
-  };
-
-  const getAddUserErrorMessage = () => {
-    if ((addUserError as { status: number }).status === 401) {
-      return t('add_altinn2_account_page.invalid_credentials');
-    } else if ((addUserError as { status: number }).status === 429) {
-      return t('add_altinn2_account_page.too_many_requests_error');
-    }
-    return t('add_altinn2_account_page.add_account_error');
-  };
 
   const step1Component = (
     <DsDialog.Block className={classes.addAltinn2Account}>
@@ -139,60 +115,15 @@ export const AddAltinn2AccountPage = () => {
   );
 
   const step2Component = (
-    <>
-      <DsDialog.Block className={classes.addAltinn2AccountHeader}>
-        <PersonCircleIcon fontSize={'1.5rem'} />
-        <DsHeading level={1}>{t('add_altinn2_account_page.add_account_heading')}</DsHeading>
-      </DsDialog.Block>
-      <DsDialog.Block className={classes.addAltinn2Account}>
-        <DsParagraph>
-          {t('add_altinn2_account_page.login_info', { name: reportee?.name })}
-        </DsParagraph>
-        <DsTextfield
-          label={t('add_altinn2_account_page.username')}
-          data-size='sm'
-          value={userName}
-          onChange={(e) => setUserName(e.target.value)}
-          onKeyDown={handleInputFieldKeyDown}
-        />
-        <DsTextfield
-          label={t('add_altinn2_account_page.password')}
-          data-size='sm'
-          type='password'
-          autoComplete='off'
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          onKeyDown={handleInputFieldKeyDown}
-        />
-        {addUserError && <DsAlert data-color='danger'>{getAddUserErrorMessage()}</DsAlert>}
-        <div className={classes.buttonRow}>
-          <DsButton
-            variant='primary'
-            aria-disabled={isValidateCrentialsButtonDisabled}
-            loading={isAddingAltinn2Account}
-            onClick={onAddAltinn2Account}
-          >
-            {t('add_altinn2_account_page.add_account_button')}
-          </DsButton>
-          <DsButton
-            variant='secondary'
-            asChild
-          >
-            <a href={afUrl}>{t('common.cancel')}</a>
-          </DsButton>
-          <DsButton
-            variant='tertiary'
-            onClick={() => {
-              setUserName('');
-              setPassword('');
-              setStep(4);
-            }}
-          >
-            {t('add_altinn2_account_page.forgot_password')}
-          </DsButton>
-        </div>
-      </DsDialog.Block>
-    </>
+    <Altinn2AccountForm
+      onAddAltinn2Account={onAddAltinn2Account}
+      isAddingAltinn2Account={isAddingAltinn2Account}
+      addUserError={addUserError}
+      onSendForgotPasswordEmail={onSendForgotPasswordEmail}
+      isSendingForgotPasswordEmail={isSendingForgotPasswordEmail}
+      forgotPasswordError={forgotPasswordError}
+      reporteeName={reportee?.name}
+    />
   );
 
   const step3Component = (
@@ -211,47 +142,6 @@ export const AddAltinn2AccountPage = () => {
           asChild
         >
           <a href={profileUrl}>{t('add_altinn2_account_page.go_to_profile')}</a>
-        </DsButton>
-      </div>
-    </DsDialog.Block>
-  );
-
-  const forgotPasswordStepComponent = (
-    <DsDialog.Block className={classes.addAltinn2Account}>
-      <DsHeading>{t('add_altinn2_account_page.forgot_password')}</DsHeading>
-      <DsParagraph>{t('add_altinn2_account_page.forgot_password_info')}</DsParagraph>
-      <DsTextfield
-        label={t('add_altinn2_account_page.username')}
-        data-size='sm'
-        value={userName}
-        onChange={(e) => setUserName(e.target.value)}
-        onKeyDown={(e: React.KeyboardEvent<HTMLInputElement>) => {
-          if (e.key === 'Enter') {
-            e.preventDefault();
-            onSendForgotPasswordEmail();
-          }
-        }}
-      />
-      {forgotPasswordError && (
-        <DsAlert data-color='danger'>{t('add_altinn2_account_page.forgot_password_error')}</DsAlert>
-      )}
-      <div className={classes.buttonRow}>
-        <DsButton
-          variant='primary'
-          aria-disabled={isSendEmailButtonDisabled}
-          loading={isSendingForgotPasswordEmail}
-          onClick={onSendForgotPasswordEmail}
-        >
-          {t('add_altinn2_account_page.send_link')}
-        </DsButton>
-        <DsButton
-          variant='secondary'
-          onClick={() => {
-            setStep(2);
-            setUserName('');
-          }}
-        >
-          {t('common.cancel')}
         </DsButton>
       </div>
     </DsDialog.Block>
@@ -301,8 +191,7 @@ export const AddAltinn2AccountPage = () => {
         {step === 1 && step1Component}
         {step === 2 && step2Component}
         {step === 3 && step3Component}
-        {step === 4 && forgotPasswordStepComponent}
-        {step === 5 && forgotPasswordReceiptComponent}
+        {step === 4 && forgotPasswordReceiptComponent}
       </DsDialog>
     </PageLayoutWrapper>
   );
